Address Unknown's feedback
https://github.com/hrydgard/ppsspp/pull/7965/files
