+++ Core/HLE/sceKernelThread.h
@@ -330,3 +330,7 @@ KernelObject *__KernelThreadEventHandlerObject();
 SceUID sceKernelRegisterThreadEventHandler(const char *name, SceUID threadID, u32 mask, u32 handlerPtr, u32 commonArg);
 int sceKernelReleaseThreadEventHandler(SceUID uid);
 int sceKernelReferThreadEventHandlerStatus(SceUID uid, u32 infoPtr);
+
+int sceKernelCreateHeap(int partitionId, int size, int flags, const char *Name);
+int sceKernelAllocHeapMemory(int heapId, int size);
+int sceKernelDeleteHeap(int heapId);
